refactor: move cohorts data to its own folder
