+++ src/data/apiHook.ts
@@ -1,6 +1,6 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { getCourseInfo, getCohorts, getCohortStatus, toggleCohorts } from './api';
-import { courseInfoQueryKeys, cohortsQueryKeys } from './queryKeys';
+import { useQuery } from '@tanstack/react-query';
+import { getCourseInfo } from './api';
+import { courseInfoQueryKeys } from './queryKeys';
 
 export const useCourseInfo = (courseId: string) => (
   useQuery({
@@ -8,29 +8,3 @@ export const useCourseInfo = (courseId: string) => (
     queryFn: () => getCourseInfo(courseId),
   })
 );
-
-export const useCohortStatus = (courseId: string) => (
-  useQuery({
-    queryKey: cohortsQueryKeys.enabled(courseId),
-    queryFn: () => getCohortStatus(courseId),
-    enabled: !!courseId,
-  })
-);
-
-export const useCohorts = (courseId: string) => (
-  useQuery({
-    queryKey: cohortsQueryKeys.list(courseId),
-    queryFn: () => getCohorts(courseId),
-    enabled: !!courseId,
-  })
-);
-
-export const useToggleCohorts = (courseId: string) => {
-  const queryClient = useQueryClient();
-  return (useMutation({
-    mutationFn: ({ isCohorted }: { isCohorted: boolean }) => toggleCohorts(courseId, isCohorted),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: cohortsQueryKeys.enabled(courseId) });
-    },
-  }));
-};
